feat:add class Booking
