--- src/main/java/ru/practicum/shareit/booking/Booking.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.booking;
-
-/**
- * TODO Sprint add-bookings.
- */
-public class Booking {
-}
