+++ src/main/java/ru/practicum/shareit/booking/model/Booking.java
@@ -1,0 +1,26 @@
+package ru.practicum.shareit.booking.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import org.springframework.stereotype.Component;
+import ru.practicum.shareit.item.model.Item;
+import ru.practicum.shareit.user.model.User;
+
+import java.time.LocalDateTime;
+
+/**
+ * TODO Sprint add-bookings.
+ */
+
+@Data
+@AllArgsConstructor
+@Component
+public class Booking {
+    private Long id;
+    private LocalDateTime start;
+    private LocalDateTime end;
+    Item item;
+    User booker;
+    Status status;
+
+}
