+++ src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.request;
+package ru.practicum.shareit.request.model;
 
 /**
  * TODO Sprint add-item-requests.
